+++ spec/api/controllers/shared/kigo_price_quotation.rb
@@ -127,9 +127,9 @@ RSpec.shared_examples "Kigo price quotation" do
       expect(response.body["check_out"]).to eq "2016-03-25"
       expect(response.body["guests"]).to eq 2
       expect(response.body["currency"]).to eq "EUR"
-      expect(response.body["net_rate"]).to eq 532.71
+      expect(response.body["net_rate"]).to eq 530.1
       expect(response.body["total"]).to eq 570.0
-      expect(response.body["host_fee"]).to eq 37.29
+      expect(response.body["host_fee"]).to eq 39.9
       expect(response.body["host_fee_percentage"]).to eq 7
     end